Activate different redis servers in resque without breaking main adapter
The way we were making this thread-safe didn't work when we run a multi-threaded server
such as Puma, because we end up configuring Resque.redis only in the main thread.
Threads in the thread pool would see Resque.redis empty, and would default to
localhost:6379. This doesn't happen in Dash right now because Dash uses Unicorn, so the
code runs in a single thread.

With this change, we keep a separate Redis instance in Resque for overriding. When it's not set,
we use the regular instance stored in `@data_store`, that gets configured usually in an
initializer and it's shared among all threads. When it's set, we use that one. This will be set
only in requests that involve Mission Control jobs and where we need to switch the adapter, but
not during regular app operation.