Limit Solid Queue count queries internally and return Infinity if above limit
For very big tables, a COUNT(*) would just take too long, and this is something
we show in almost every page in the navigation. We don't really need it, though,
since we can paginate and fetch rows with a small limit and an offset, we don't
need to know how many pages there are. Take advantage of that.