Return last Solid Queue job when there are several with the same Active Job ID
This applies in the case of automatic retries performed by Active Job's
retrying mechanism. In that case, the failing job finishes just fine,
the exception is retried and Active Job re-enqueues the job, so for
Solid Queue, these look like different jobs with the same Active Job ID.
Normally we'd be interested in the last one only, though, so just return
that. In the future perhaps we can find a way to return all the instances
if there are more.