Fix workers pagination issue where the whole relation was being loaded
Due to the delegation of all count-related methods to `to_a`, we were loading
the whole set of workers when calculating the number of pages and page size,
which happens before we've set any limits or offsets in the relation. Moreover,
once the limits and offsets are set, we weren't resetting the loaded workers, which
means we were still returning all of them even with offset and limit.

This adds also a test for this and includes some minor cosmetic fixes.