Fixes initializing the Rails.backtrace_cleaner in the Gem Vs. App
I inadvertantly early initialized the Rails.backtrace_cleaner as part
of setting MissionControl wide defaults.  This commit uses a proc
as the default so the database cleaner will be initialized in the
Application environment (vs. the Gems's)

Fixed a typo where I referenced the Gems state instead of the Applicaations
state to access the database cleaner