Fix "expires in" message in blocked jobs table
The `bidirectional_time_distance_in_words_with_title` helper outputs a html-safe span tag. However it is used in string interpolation in the blocked jobs table making which is not html-safe by default resulting in the the raw html being displayed on the page.

This calls `#html_safe` on the string to not escape the markup.